refactor: Move the tests of combinator.rs to tests/
No reason to have these in the file itself.
